+++ test/integrations/expocontext.test.ts
@@ -1,23 +1,11 @@
-import type { Hub } from '@sentry/core';
-import type { Event } from '@sentry/types';
+import type { Client, Event } from '@sentry/types';
 
-import { ExpoContext } from '../../src/js/integrations/expocontext';
+import { expoContextIntegration } from '../../src/js/integrations/expocontext';
 import { getExpoDevice } from '../../src/js/utils/expomodules';
 
 jest.mock('../../src/js/utils/expomodules');
 
 describe('Expo Context Integration', () => {
-  let integration: ExpoContext;
-
-  const mockGetCurrentHub = () =>
-    ({
-      getIntegration: () => integration,
-    } as unknown as Hub);
-
-  beforeEach(() => {
-    integration = new ExpoContext();
-  });
-
   it('does not add device context because expo device module is not available', async () => {
     (getExpoDevice as jest.Mock).mockReturnValue(undefined);
     const actualEvent = await executeIntegrationFor({});
@@ -112,16 +100,7 @@ describe('Expo Context Integration', () => {
     });
   });
 
-  function executeIntegrationFor(mockedEvent: Event): Promise<Event | null> {
-    return new Promise((resolve, reject) => {
-      integration.setupOnce(async eventProcessor => {
-        try {
-          const processedEvent = await eventProcessor(mockedEvent, {});
-          resolve(processedEvent);
-        } catch (e) {
-          reject(e);
-        }
-      }, mockGetCurrentHub);
-    });
+  function executeIntegrationFor(mockedEvent: Event): Event {
+    return expoContextIntegration().processEvent!(mockedEvent, {}, {} as Client) as Event;
   }
 });